+++ packages/core/src/components/segmented-button/segmented-button.interfaces.ts
@@ -1,0 +1,9 @@
+export interface SegmentedButtonOption {
+  label: string;
+  disabled?: boolean;
+}
+
+export interface SegmentedButtonChangeEvent {
+  originalEvent: Event;
+  option: number;
+}