+++ packages/dso-toolkit/src/components/segmented-button/segmented-button.models.ts
@@ -1,0 +1,14 @@
+export interface SegmentedButtonChangeEvent {
+  originalEvent: Event;
+  option: number;
+}
+export interface SegmentedButton {
+  options: SegmentedButtonOption[];
+  activeOption: number;
+  dsoChange?: (event: CustomEvent<SegmentedButtonChangeEvent>) => void;
+}
+
+export interface SegmentedButtonOption {
+  label: string;
+  disabled?: boolean;
+}